+++ HISTORY.rst
@@ -12,6 +12,7 @@ Added
     - `level.exposure_level` for the sound exposure level (PR #951)
     - `level.sliding_continuous_level` for the Leq with a sliding window (PR #952)
     - `level.peak_level` for (true) peak level detection (PR #955)
+    - `level.time_weighted_level` and `level.time_weighted_pressure` for time-weighting (PR #954)
 
 Fixed
 ^^^^^